Files structure changed and adding logging feature
