+++ src/error-handlers/error.js
@@ -1,6 +1,7 @@
 import ErrorMessages from "./error-messages";
 
-export function ParentNotFound(res){
+export function ParentNotFound(route,res){
+  log.error(`Route : '${route}' is not declared...`)
   res.writeHead(404, {"Content-Type": "application/json"});
   res.write(ErrorMessages.PARENT_NOT_FOUND);
   res.end();
